requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
motor==3.7.1
pymongo==4.16.0
